# generated/netatmo/server/nodejs/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Netatmo",
    "description": "<h3>Welcome to the Netatmo swagger on-line documentation !</h3>This site is a complement to the official <a href=\"https://dev.netatmo.com/\">Netatmo developper documentation</a> using swagger to bring interactivity and easy testing of requests with the \"try it\" button (authenticate with the authorization code 0Auth2 flow by clicking the authenticate button in the methods). You can find the source code for this site can be found in the project <a href=\"https://github.com/cbornet/netatmo-swagger-ui\">netatmo-swagger-ui</a>. You can also use the online <a href=\"./swagger.json\">swagger declaration</a> file to generate code or static documentation (see <a href=\"https://github.com/cbornet/netatmo-swagger-api\">netatmo-swagger-api</a>).",
    "version": "1.0.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
  "/api/setthermpoint": {
    
    "post": {
      "summary": "",
      "description":"The method setthermpoint changes the Thermostat manual temperature setpoint.",
      "x-swagger-router-controller": "Thermostat",
      "tags": ["Thermostat"],
      "operationId": "setthermpoint",
      "parameters": [
        {
  "name" : "access_token",
  "in" : "query",
  "description" : "Your access token",
  "required" : true,
  "type" : "string"
},
        {
  "name" : "device_id",
  "in" : "query",
  "description" : "The relay id",
  "required" : true,
  "type" : "string"
},
        {
  "name" : "module_id",
  "in" : "query",
  "description" : "The thermostat id",
  "required" : true,
  "type" : "string"
},
        {
  "name" : "setpoint_mode",
  "in" : "query",
  "description" : "Chosen setpoint_mode",
  "required" : true,
  "type" : "string",
  "enum" : [ "program", "away", "hg", "manual", "off", "max" ]
},
        {
  "name" : "setpoint_endtime",
  "in" : "query",
  "description" : "When using the manual or max setpoint_mode, this parameter defines when the setpoint expires.",
  "required" : false,
  "type" : "integer",
  "format" : "int32"
},
        {
  "name" : "setpoint_temp",
  "in" : "query",
  "description" : "When using the manual setpoint_mode, this parameter defines the temperature setpoint (in Celcius) to use.",
  "required" : false,
  "type" : "number",
  "format" : "float"
}
        
      ],
      "responses": {
        "200": {
  "description" : "Successful response",
  "schema" : {
    "$ref" : "#/definitions/NAOkResponse"
  }
}
        
        
      }
    } 
    
  } ,
  "/api/createnewschedule": {
    
    "post": {
      "summary": "",
      "description":"The method createnewschedule creates a new schedule stored in the backup list.",
      "x-swagger-router-controller": "Thermostat",
      "tags": ["Thermostat"],
      "operationId": "createnewschedule",
      "parameters": [
        {
  "name" : "access_token",
  "in" : "query",
  "description" : "Your access token",
  "required" : true,
  "type" : "string"
},
        {
  "name" : "device_id",
  "in" : "query",
  "description" : "The relay id",
  "required" : true,
  "type" : "string"
},
        {
  "name" : "module_id",
  "in" : "query",
  "description" : "The thermostat id",
  "required" : true,
  "type" : "string"
},
        {
  "in" : "body",
  "name" : "therm_program",
  "description" : "The thermostat program (zones and timetable)",
  "required" : true,
  "schema" : {
    "$ref" : "#/definitions/NAThermProgram"
  }
}
        
      ],
      "responses": {
        "200": {
  "description" : "Successful response",
  "schema" : {
    "$ref" : "#/definitions/NANewScheduleResponse"
  }
}
        
        
      }
    } 
    
  } ,
...
  }, "definitions": {
...
},"NAResponse": {
  "properties" : {
    "status" : {
      "type" : "string"
    },
    "time_exec" : {
      "type" : "number",
      "format" : "double"
    },
    "time_server" : {
      "type" : "integer",
      "format" : "int32"
    }
  }
},"NAOkResponse": {
  "allOf" : [ {
    "$ref" : "#/definitions/NAResponse"
  }, {
    "type" : "object",
    "properties" : {
      "body" : {
        "type" : "string"
      }
    }
  } ]
},"NAUserResponse": {
...
},"NANewScheduleResponse": {
  "allOf" : [ {
    "$ref" : "#/definitions/NAResponse"
  }, {
    "type" : "object",
    "properties" : {
      "body" : {
        "$ref" : "#/definitions/NANewScheduleBody"
      }
    }
  } ]
},"NAPartnerDevicesResponse": {
...
},"NANewScheduleBody": {
  "properties" : {
    "schedule_id" : {
      "type" : "string",
      "description" : "Id of the schedule. It can be use with the method \"switchschedule\" \n"
    }
  }
},"NAUserAdministrative": {
...
},"NAThermProgram": {
  "properties" : {
    "program_id" : {
      "type" : "string"
    },
    "name" : {
      "type" : "string"
    },
    "zones" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/NAZone"
      }
    },
    "timetable" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/NATimeTableItem"
      }
    },
    "selected" : {
      "type" : "boolean",
      "default" : false
    }
  }
},"NAZone": {
  "properties" : {
    "id" : {
      "type" : "integer",
      "format" : "int32"
    },
    "type" : {
      "type" : "integer",
      "format" : "int32"
    },
    "name" : {
      "type" : "string"
    },
    "temp" : {
      "type" : "number",
      "format" : "float"
    }
  }
},"NATimeTableItem": {
  "properties" : {
    "id" : {
      "type" : "integer",
      "format" : "int32"
    },
    "m_offset" : {
      "type" : "integer",
      "format" : "int32"
    }
  }
},"NAHeatingSystem": {
...
  }
}